Hillclimb

+++ src/main/java/Algorithms/Metrics.java
@@ -4,14 +4,14 @@ import java.util.ArrayList;
 
 public class Metrics {
     private int score;
-    private int timeAchieved;
+    private double timeAchieved;
     private ArrayList<ArrayList<String>> map;
 
-    public Metrics(int score, int timeAchieved, ArrayList<ArrayList<String>> map) {
+    public Metrics(int score, double timeAchieved, ArrayList<ArrayList<String>> map) {
         this.score = score;
         this.timeAchieved = timeAchieved;
         this.map = map;
-    }
+}
 
     public ArrayList<ArrayList<String>> getMap() {
         return map;
@@ -21,7 +21,7 @@ public class Metrics {
         return score;
     }
 
-    public int getTimeAchieved() {
+    public double getTimeAchieved() {
         return timeAchieved;
     }
 }
